services/model/requirements.txt
requests
yfinance
pandas
numpy
scikit-learn
kafka-python
optuna